Basic Presto Testing using Github Actions (#16)
Added a basic github action to test the presto builds with the current
testing scripts.

Currently this provides options to build `java`, `native_cpu`, and
`native_gpu` systems, and then run the basic test suite.

There is no build caching, or other complexity, set up, so the builds
are rather slow (especially for the native-gpu build).