Velox benchmarking fixups and improvements (#23)
Improvements and fixes of Velox benchmark script

* https://github.com/rapidsai/velox-testing/issues/20 - remove the
`--data-dir` CLI option and replace it with `--tpch-data-dir` which
allows specifying the dataset location directly without requiring a
certain structure. This also allows easy extension to other future
datasets such as `tpc-ds` using the same pattern.

* https://github.com/rapidsai/velox-testing/issues/19 - update the
wording and help message to make this clearer now

* Fix an issue with TPC-H help message printing `basename` instead of
script name

* Use docker env files to pass variables that are necessary to
`docker-compose`

* Detect whether the TPC-H data dir layout is in managed or external
Hive table layout. Notify user we do not currently support external Hive
table layout. If this is a use-case we want to support we can add
support for it later.

* In the lastest upstream cudf-velox branch, the CPU benchmark
executable no longer supports GPU-related options, update the script to
not pass these along.

* Enable collection of GPU metrics in `nsys`. Default to `all` GPUs.

* Pass the user and group id of current user to `docker compose` so that
output files on host can be updated to have the same ownership, as
opposed to be root-owned.

* Ensure `/opt/velox-build/release/_deps/rmm-build` is in the system
library link path to allow resolving this (seems to be needed in newer
upstream velox builds.)

* https://github.com/rapidsai/velox-testing/issues/21: partially
addresses this item by setting the default `num_repeats` to `2` and
exposing it as a parameter.

* Default `CUDA_ARCHITECTURES` to `NATIVE`