Added licensing (#42)
Added licensing required for OSRB.

- Added LICENSE file for Apache license.
- Added CONTRIBUTING.md file. This is somewhat sparse for now, with a
few TODOs; we can fill this out in more detail later once the process
has been fleshed-out.
- Added licensing headers to *most* source files. These were added to
any `.sh` or `.py` files of interest (I skipped trivial files that only
had a single line or two). I also did not add this to any `yaml` files,
as that seems to not be common for our other projects.