Moved duckdb schema calls outside of multi-threaded portions (#50)
The rewrite_parquet portions of the data generation were making
concurrent duckdb calls on the same connection, which is not
thread-safe.

This patch runs all the duckdb calls sequentially before the
multi-threaded portion of the code, storing the results in a dictionary
that can be accessed in a thread-safe way.