Pull Presto deps container image from S3 and enable Presto CI (#74)
This PR implements the ability to pull the Presto/Velox dependencies
container image from an S3 bucket instead of having to build it locally
if it does not already exist.

This is faster, especially for CI jobs.

Note that valid AWS access keys and bucket info must be set in your
environment in order for the deps container image fetch to work.

This PR also activates the nightly Presto CI job. For now, the test step
just runs our Integration Tests (TPCH SF 0.01) but disabling Q15 on GPU
since that still regularly fails due to a schema/comparison mismatch
(this is being resolved separately).