Default Presto GPU builds to native CUDA architecture (#84)
We will always want the GPU build to be optimized for the machine it
will ultimately run on, which for CI and bench-marking will be the same
machine it builds on.

We determine the native CUDA architecture using `nvidia-smi` in the
start script and pass it into the Docker Compose Build process, because
despite the given `service` having `runtime: nvidia`, that only applies
to the final container and not the build context, which therefore does
not have access to the GPU to allow `nvidia-smi` or `nvcc -arch=native`
to be used.

The Dockerfile default is still to build for all architectures, if the
`CUDA_ARCHITECTURES` ARG is not passed in.

Building for just one architecture reduces the CI build time for the GPU
variant from ~4 hours down to ~1.5 hours.