New Presto Deps Container Build Script (#80)
This updates the Presto dependencies/run-time container image build process, to
match the pre-built images already in use.

The critical part is that we override the scripts and CMake config in the Presto Velox
sub-module to be those from a sibling Velox clone. Those are the only parts of the
Velox sub-module that are actually needed for a build, in order that the container
includes the current set of SDKs required for a Velox build.

The script simple replaces just those two sub-directories with those from the sibling
Velox clone, whatever that may be. After the build is done, it undoes that change.

This process may require updating if the build relies on more Velox components in
the future, but this seems unlikely. In fact, the recent incompatibility was really
only due to the update in CUDA libraries needed to support cuDF 25.10, and such
an update may not be required again for a while, so once those changes land in
upstream Velox, we may not need to do the sub-directory substitution at all.