Fail Presto CPU or GPU build if no deps image is found (#88)
This PR changes the default behavior of the start_native_[cpu|gpu]_presto.sh scripts.

If the corresponding worker Docker container image is not found, it will automatically
be built, but only if the Presto dependencies/run-time base image is available.
Previously, this too would be built automatically, but this is no longer appropriate now
that we have the option of fetching a pre-built image.

The `build_centos_deps_image.sh` or `fetch_centos_deps_image.sh` script must be run
manually in this case. Use of the latter may require additional credentials.

Includes corresponding update to `README.md`.