Generate dynamic Presto configs with pbench (#48)
This PR adds the ability to generate Presto config values dynamically for a
given machine spec (CPU count and CPU RAM size).

It adds a script which calls pbench genconfig to generate the config files,
which is itself called from the start scripts so that a valid config exists before
the server starts. The config files are then free files based on templates.

Note that the Java variant still uses different config files which are hard-wired
and do not contain dynamic values. These remain at the basic default settings.

Note also that there is variant-specific behavior in that `vcpu_per_worker` is
set to 2 for GPU mode, and to CPU count for CPU and Java.

---------

Co-authored-by: Simon Eves <seves@nvidia.com>
Co-authored-by: misiugodfrey <misiug@nvidia.com>