Simplify Velox build changes (#90)
The changes that landed in #64 are more complex than was required. This
PR simplifies them, and brings the mechanism more in line with the one
for Presto that landed last week.

Instead of creating an intermediate Docker image, we simply leverage the
existing mechanism that is used to generate the upstream base image
(ghcr.io/facebookincubator/velox-dev:adapters) to build that image locally
using the modified Velox build scripts which now include nvjitlink package
install.

This mirrors what we expect Facebook to do themselves once our cuDF
25.10 PR lands in upstream. Once that is done, we will then have the
option to revert to using the upstream image.

Also not reverted is the removal of the --pull option from the Velox
Docker build, as otherwise the build would always pull the upstream
version of that image instead of using our modified local version.

Finally, the build and fetch scripts are adapted accordingly.