Add option to Presto CI to set VELOX_ENABLE_BACKWARD_COMPATIBLE in Velox compile (#94)
This is a useful option which allows Presto compiles to work in the case where there
has been a Velox API change and hence a temporary mismatch, in which case the
Velox code usually has `#ifdef VELOX_ENABLE_BACKWARD_COMPATIBILITY` logic to 
temporarily retain the previous API. This isn't always the case, but it usually is.

As of today (10/23) there is a mismatch which requires this option. This PR adds a CI
job option to conditionally patch the Presto Velox CMake config to add that compile
definition, which hopefully avoids a CI blockage while the mismatch is resolved.