Added casting, and error checking for pbench config script (#98)
The pbench config script was swallowing errors such that it was not
clear when it failed (and the presto start scripts would continue even
if it had, using/creating invalid config files). With this patch we cast
mem values to int (the original failure was from floats being passed in
and failing parsing), we check the pbench output for errors, and will
have the script return an error value if the config failed (thereby
halting the start* scripts if they call it and fail).