Extend Presto worker start-up time-out to 60s (#99)
The existing time-out is 25s, which isn't long enough as workers take up to ~30s to start.

This PR increases the retry count from 5 to 12, this increasting the the timeout to 60s.

This was prompted by various intermittent failures in Presto CI whereby the Integration
Tests fail because the worker doesn't start quickly enough.