Presto: More robust handling of existing config (#109)
This PR adds logic to the Presto start scripts to re-use any existing config files for the
given variant in case they have been modified by the user.

Config files for the given variant will be automatically generated if absent.

A config tree is generated for each variant independently, so that the user is free to
make manual edits in the generated version for one variant, but switch to another
variant temporarily (e.g. to do an ANALYZE in CPU mode) without losing the edits.

Also added an `--overwrite-config` option to the start scripts to force regeneration
of the config when deliberately switching variants or to revert an edited config back
to default.

Configs generated or edited with previous versions of velox-testing will be ignored,
and should ideally be removed manually