Generate duckdb schemas from parquet samples instead of entire files (#116)
Quick fix to have our schema generation scripts use less memory.

Currently, we extract the schema from parquet files by using them to
create a duckdb table and then do a `SHOW CREATE TABLE` on the duckdb
table to get an SQL representation. This also lets us alter the table
schema if we need to while it exists in duckdb (we do this to apply NOT
NULL to columns, which do not exist in the parquet schema). This fix
adds a limit on the duckdb CTAS used in this case so that we are not
generating a gigantic table when using large parquet files, as we only
need a small data sample to know the schema.