Presto: Fix Java Automatic Memory Config (#117)
Previous attempts at allowing automatic memory config for Java mode
failed, with the Coordinator failing to start, with errors of the form:

  Invalid memory configuration. The sum of max total query memory per
  node (96636764160) and heap headroom (36399847833) cannot be larger
  than the available heap memory (121332826112)

This turns out to be because there are two additional memory-related
configs that were in the Native Coordinator config but not in the Java one
which turn out to be essential in order for the memory assignment math
to be correct.

Thanks to @paul-aiyedun for finding the fix.

Also fixes the Java config.properties file container mapping broken
accidentally in #109.