Presto: Add Nightly (Upstream) CI (#120)
This PR adds a second nightly job which builds Presto with their default
pinned upstream Velox instead of ours.

This is achieved by having a pre-job which fetches the pinned version SHA
from the Presto sub-module, so that we then check out that same SHA for
the sibling Velox clone. This allows the use of our normal build recipe which
expects Velox to be in the sibling clone not the sub-module. While this job
is arguably an expensive resource, it only takes about 20 seconds to run.

This PR also removes the Velox Testing Commit option and associated
plumbing from the Presto workflow stack. This was intended to allow devs
to run a workflow from a branch and specify that same branch for the
checkout of velox-testing so that any script additions or changed matched
the workflow being tested. Turns out there's a pre-defined GHA variable
which contains the branch name that the workflow is being run from, so
we can use that directly for the velox-testing checkout so that they match
automatically.