Fix Presto native worker build fail due to nsight-systems-install fail (#133)
The package name for that version of `nsight-systems` appears to have changed in the
last day or so. It is no longer installable with the full version number string. Removing
the final .121 suffix resolves the issue, but we remove the version number entirely so
as to always get the latest version (which is now 2025.6.1.190).