Presto: Separate Hive configs for Coordinator & Worker (#121)
There are additional Hive config options which can be used in the Worker
but if they are also in the Coordinator, it throws an error.

For now, we just add some Parquet read parameters at their defaults, and
is therefore just a convenience for users who wish to tweak these for
some specific machine.

**IMPORTANT!**
You will need to use `--overwrite-config` when first running with this
PR, otherwise it will reuse the existing tree, and there will be a
`hive.properties` file in the old location which will throw a startup
error in the Docker mappings. Be sure to copy any edited files aside
before doing this, of course, or they will be lost.